Mise à jour du projet avec la methode de trie de calcul de marge et l'intérgation de l'exceptio pour les produits avec prix d'achat inférieur ou égale à 0

+++ src/main/java/com/ecommerce/microcommerce/web/controller/ProductController.java
@@ -1,22 +1,35 @@
 package com.ecommerce.microcommerce.web.controller;
 
+import java.net.URI;
+import java.util.Iterator;
+import java.util.List;
+
+import javax.validation.Valid;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.json.MappingJacksonValue;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
 import com.ecommerce.microcommerce.dao.ProductDao;
 import com.ecommerce.microcommerce.model.Product;
+import com.ecommerce.microcommerce.web.exceptions.ProduitGratuitException;
 import com.ecommerce.microcommerce.web.exceptions.ProduitIntrouvableException;
 import com.fasterxml.jackson.databind.ser.FilterProvider;
 import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
 import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
+
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.http.converter.json.MappingJacksonValue;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import javax.validation.Valid;
-import java.net.URI;
-import java.util.List;
 
 
 @Api( description="API pour es opérations CRUD sur les produits.")
@@ -36,7 +49,7 @@ public class ProductController {
 
         Iterable<Product> produits = productDao.findAll();
 
-        SimpleBeanPropertyFilter monFiltre = SimpleBeanPropertyFilter.serializeAllExcept("prixAchat");
+        SimpleBeanPropertyFilter monFiltre = SimpleBeanPropertyFilter.serializeAllExcept("marge");
 
         FilterProvider listDeNosFiltres = new SimpleFilterProvider().addFilter("monFiltreDynamique", monFiltre);
 
@@ -52,13 +65,22 @@ public class ProductController {
     @ApiOperation(value = "Récupère un produit grâce à son ID à condition que celui-ci soit en stock!")
     @GetMapping(value = "/Produits/{id}")
 
-    public Product afficherUnProduit(@PathVariable int id) {
+    public MappingJacksonValue afficherUnProduit(@PathVariable int id) {
 
         Product produit = productDao.findById(id);
 
         if(produit==null) throw new ProduitIntrouvableException("Le produit avec l'id " + id + " est INTROUVABLE. Écran Bleu si je pouvais.");
+        
+        
+        SimpleBeanPropertyFilter monFiltre = SimpleBeanPropertyFilter.serializeAllExcept("marge");
 
-        return produit;
+        FilterProvider listDeNosFiltres = new SimpleFilterProvider().addFilter("monFiltreDynamique", monFiltre);
+
+        MappingJacksonValue produitsFiltres = new MappingJacksonValue(produit);
+
+        produitsFiltres.setFilters(listDeNosFiltres);
+
+        return produitsFiltres;
     }
 
 
@@ -68,8 +90,11 @@ public class ProductController {
     @PostMapping(value = "/Produits")
 
     public ResponseEntity<Void> ajouterProduit(@Valid @RequestBody Product product) {
-
-        Product productAdded =  productDao.save(product);
+    	
+    	
+    	if((product==null) || (product.getPrixAchat()<=0)) throw new ProduitGratuitException("Le produit que vous voulez enregistrer possède un prix d'achat inférieur ou égale à zéro.");
+        
+    	Product productAdded =  productDao.save(product);
 
         if (productAdded == null)
             return ResponseEntity.noContent().build();
@@ -84,9 +109,17 @@ public class ProductController {
     }
 
     @DeleteMapping (value = "/Produits/{id}")
-    public void supprimerProduit(@PathVariable int id) {
+    public ResponseEntity<Void> supprimerProduit(@PathVariable int id) {
 
-        productDao.delete(id);
+    	Product product=productDao.findById(id);
+		
+		if(product== null){
+			return ResponseEntity.notFound().build();
+		}
+		productDao.delete(id);
+	      
+
+		return ResponseEntity.ok().build();
     }
 
     @PutMapping (value = "/Produits")
@@ -104,5 +137,50 @@ public class ProductController {
     }
 
 
+
+    @ApiOperation(value = "Affiche les produits avec leurs marges beneficiaire")
+    @GetMapping(value = "/AdminProduits")
+    public MappingJacksonValue calculerMargeProduit(){
+    	
+    	List<Product> produits = productDao.findAll();
+    	 
+    	    	 
+    	Iterator itProduct = produits.iterator();
+    	 while (itProduct.hasNext()) {
+			Product produit = (Product) itProduct.next();
+			 produit.setMarge(produit.getPrix()-produit.getPrixAchat());
+		}   	   	 
+    	 
+    	 
+    	 SimpleBeanPropertyFilter monFiltre = SimpleBeanPropertyFilter.serializeAll();
+
+         FilterProvider listDeNosFiltres = new SimpleFilterProvider().addFilter("monFiltreDynamique", monFiltre);
+
+         MappingJacksonValue produitsFiltres = new MappingJacksonValue(produits);
+
+         produitsFiltres.setFilters(listDeNosFiltres);
+    	
+    	return produitsFiltres;
+    	
+    }
+    
+    @ApiOperation(value = "Affiche les produits triés par ordre aphabetique des noms!")
+    @GetMapping (value = "/TrieProduits")
+    public MappingJacksonValue  trierProduitsParOrdreAlphabetique(){
+    	
+    	 Iterable<Product> produits = productDao.trieProduits();;
+
+         SimpleBeanPropertyFilter monFiltre = SimpleBeanPropertyFilter.serializeAllExcept("marge");
+
+         FilterProvider listDeNosFiltres = new SimpleFilterProvider().addFilter("monFiltreDynamique", monFiltre);
+
+         MappingJacksonValue produitsFiltres = new MappingJacksonValue(produits);
+
+         produitsFiltres.setFilters(listDeNosFiltres);
+
+         return produitsFiltres;
+    	
+    }
+    
 
 }
